feat: Add WIP delete relations button.
This will be used to present a modal by which the user can select filters for which CSP violations to delete.
This provides a robust way to delete only the reports a user might want deleted (e.g. all blocking a specific URI/host, or on a specific effective directive, or before a certain date) without making assumptions about what those options are.