feat: Add SecurityHeaderMiddleware.
Because of the way we tend to use this module (i.e. apply it to the base controller class), it makes more sense to use middleware than a controller. This ensures that even responses that are returned but not attached to the controller correctly will have headers applied.
This will also make more sense for splitting front-end and back-end CSP configurations.