fix: Check that CMSReportingOnly exists before access. SIGMAS-172
CWP deployments create the codebase, then run flush, and then run
/dev/build.

Silverstripe assumes that SiteConfig is up-to-date when
SilverStripe\SiteConfig\SiteConfig::current_site_config() is called from
Signify\Middleware\SecurityHeaderMiddleware::isCSPReportingOnly(). If it
isn't, an error is thrown.

To avoid this, verify that SiteConfig has the field before trying to
access it.