fix: Explicitly check if SiteConfig's table is ready.
ClassInfo::dataClassesFor() should respect the Injector so any injected
class that's replacing SiteConfig should be caught by this, and it is
only meant to protect against SiteConfig::current_site_config() so even
if the extension is applied to some other class, that won't matter.

This ensures that the SiteConfig table explicitly is checked.